+++ lib/web/autoupdate_rfd109.go
@@ -77,7 +77,7 @@ func (h *Handler) automaticUpgradesVersion109(w http.ResponseWriter, r *http.Req
 	ctx, cancel := context.WithTimeout(r.Context(), defaultChannelTimeout)
 	defer cancel()
 
-	targetVersion, err := h.autoUpdateAgentVersion(ctx, channelName, "" /* updater UUID */)
+	targetVersion, err := h.autoUpdateResolver.GetVersion(ctx, channelName, "" /* updater UUID */)
 	if err != nil {
 		// If the error is that the upstream channel has no version
 		// We gracefully handle by serving "none"
@@ -103,7 +103,7 @@ func (h *Handler) automaticUpgradesCritical109(w http.ResponseWriter, r *http.Re
 
 	// RFD109 agents already retrieve maintenance windows from the CMC, no need to
 	// do a maintenance window lookup for them.
-	critical, err := h.autoUpdateAgentShouldUpdate(ctx, channelName, "" /* updater UUID */, false /* window lookup */)
+	critical, err := h.autoUpdateResolver.ShouldUpdate(ctx, channelName, "" /* updater UUID */, false /* window lookup */)
 	if err != nil {
 		return nil, trace.Wrap(err)
 	}
